third_party/blink/web_tests/wpt_internal/app-history/navigate-event/navigate-iframe-location.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="iframe" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    appHistory.onnavigate = t.step_func_done(() => {
      assert_unreached("onnavigate should not have fired in source window");
    });
    iframe.contentWindow.appHistory.onnavigate = t.step_func_done(e => {
      assert_true(e.cancelable);
      assert_true(e.canRespond);
      assert_false(e.userInitiated);
      assert_true(e.hashChange);
      assert_equals(new URL(e.destination.url).hash, "#1");
      assert_true(e.destination.sameDocument);
      assert_equals(e.formData, null);
      e.preventDefault();
      t.step_timeout(t.step_func_done(() => assert_equals(location.hash, "")), 0);
    });
    iframe.contentWindow.location.hash = "#1";
  });
}, "location API on another window fires navigate event in the target window only");
</script>
